success

+++ backlar/app/Http/Controllers/Api/AuthController.php
@@ -40,18 +40,30 @@ $token = $user->createToken('main')->plainTextToken;
 return response(compact('user', 'token'));
    }
 
-   public function logout(Request $request)
-   {
-       $user = Auth::user();
-       $user->currentAccessToken()->delete();
+//    public function logout(Request $request)
+//    {
+//        $user = Auth::user();
+//        $user->currentAccessToken()->delete();
 
-       return response()->json(['message' => 'Logged out successfully.'], 200);
-   }
+//        return response()->json(['message' => 'Logged out successfully.'], 200);
 
 
-//    public function logout (Request $request) {
-// $user = $request->user();
+
+   public function logout (Request $request) {
+$user = $request->user();
+
+
+    if (!$user) {
+
+        return response(401);
+
+    }
+
+    $user->currentAccessToken()->delete();
+
+    return response('', 204);
+
 // $user -> currentAccessToken()->delete();
 // return response("", 204);
-//    }
+   }
 }
